requirements.txt
# Core dependencies
python-dotenv>=0.19.0
requests>=2.26.0
twitchio>=2.6.0

# AI integration
openai>=0.27.0
anthropic>=0.5.0

# Async utilities
aiohttp>=3.8.1
asyncio>=3.4.3

# Testing
pytest>=7.0.0
pytest-asyncio>=0.18.0

# Development tools
black>=22.1.0
flake8>=4.0.1